Revert "change to long"
This reverts commit cf6b6cb6cf739d3a354f7da662ab427e986d7626.

+++ src/main/java/io/github/levtey/TokenAPI/Place.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 public class Place {
 	
 	public final UUID uuid;
-	public final long amount;
+	public final int amount;
 	
-	public Place(UUID uuid, long amount) {
+	public Place(UUID uuid, int amount) {
 		this.uuid = uuid;
 		this.amount = amount;
 	}
